+++ tests/test_end_to_end_run.py
@@ -22,7 +22,8 @@ def set_up_s3(mocked_s3, test_folder, config):
     ]
     for b in buckets:
         mocked_s3.meta.client.create_bucket(
-            Bucket=b, CreateBucketConfiguration={"LocationConstraint": "eu-west-1"},
+            Bucket=b,
+            CreateBucketConfiguration={"LocationConstraint": "eu-west-1"},
         )
 
     files = [
@@ -44,6 +45,20 @@ def test_end_to_end(s3):
     set_up_s3(s3, test_folder, config)
     run_validation(config_path)
     os.system(f"python data_linter/command_line.py --config-path {config_path}")
+
+
+@pytest.mark.parametrize("validator", ["frictionless", "great-expectations"])
+def test_end_to_end_ge(s3, validator):
+
+    from data_linter.validation import run_validation
+
+    test_folder = "tests/data/end_to_end1/"
+    config_path = os.path.join(test_folder, "config.yaml")
+    with open(config_path) as f:
+        config = yaml.safe_load(f)
+    config["validator-engine"] = validator
+    set_up_s3(s3, test_folder, config)
+    run_validation(config)
 
 
 def test_end_to_end_no_creds_error():