+++ tests/test_headers.py
@@ -2,9 +2,8 @@ import os
 import json
 import pytest
 
-from data_linter.validation import (
-    _read_data_and_validate,
-    convert_meta_to_goodtables_schema,
+from data_linter.validators.frictionless_validator import (
+    FrictionlessValidator,
 )
 
 
@@ -23,15 +22,12 @@ from data_linter.validation import (
 def test_headers(file_name, expected_result):
     """
     Tests files against the _read_data_and_validate function.
-
     runs each file and corresponding meta (table1 or table2).
     Against the additional table config params:
     - expected-headers is False
     - expected-headers is True and ignore-case is False
     - expected-headers is True and ignore-case is True
-
     In that order
-
     Args:
         file_name ([str]): The filename in the dir tests/data/headers/
         expected_results ([Tuple(bool)]): expected results for the 3
@@ -44,8 +40,6 @@ def test_headers(file_name, expected_result):
     with open(os.path.join(test_folder, f"meta_data/{table_name}.json")) as f:
         metadata = json.load(f)
 
-    schema = convert_meta_to_goodtables_schema(metadata)
-
     table_params = [
         {"expect-header": False},
         {"expect-header": True, "headers-ignore-case": False},
@@ -54,10 +48,9 @@ def test_headers(file_name, expected_result):
 
     all_tests = []
     for table_param in table_params:
-        response = _read_data_and_validate(
-            full_file_path, schema, table_param, metadata
-        )
-        table_response = response["tables"][0]
+        validator = FrictionlessValidator(full_file_path, table_param, metadata)
+        validator.read_data_and_validate()
+        table_response = validator.response
         all_tests.append(table_response["valid"])
 
     assert expected_result == all_tests